generic/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
websockets==14.1
pydantic==2.11.9
python-dotenv==1.0.1
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
openai==1.51.0
tiktoken==0.8.0
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.27.2
mcp==1.13.0
pyyaml==6.0.3